+++ src/test/java/ampcontrol/model/training/data/iterators/preprocs/CnnToRnnToLastStepToFfPreProcessorTest.java
@@ -19,10 +19,10 @@ public class CnnToRnnToLastStepToFfPreProcessorTest {
      */
     @Test
     public void preProcess() {
-        final int[] featureShapeIn = {3,1,7,11};
-        final int[] expectedFeatureShape = {3,11,7};
-        final int[] labelsShapeIn = {3,5};
-        final int[] expectedLabelsShape = {3,5};
+        final long[] featureShapeIn = {3,1,7,11};
+        final long[] expectedFeatureShape = {3,11,7};
+        final long[] labelsShapeIn = {3,5};
+        final long[] expectedLabelsShape = {3,5};
         final DataSet testSet = new org.nd4j.linalg.dataset.DataSet(Nd4j.create(featureShapeIn), Nd4j.create(labelsShapeIn));
         final DataSetPreProcessor pp = new CnnToRnnToLastStepToFfPreProcessor();
         pp.preProcess(testSet);
